+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/TLPeerUser.java
@@ -8,6 +8,7 @@ import java.io.OutputStream;
 
 import static com.github.badoualy.telegram.tl.StreamUtils.readInt;
 import static com.github.badoualy.telegram.tl.StreamUtils.writeInt;
+import static com.github.badoualy.telegram.tl.StreamUtils.writeLong;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32;
 
@@ -19,7 +20,7 @@ public class TLPeerUser extends TLAbsPeer {
 
     public static final int CONSTRUCTOR_ID = 0x9db1bc6d;
 
-    protected int userId;
+    protected long userId;
 
     private final String _constructor = "peerUser#9db1bc6d";
 
@@ -32,7 +33,7 @@ public class TLPeerUser extends TLAbsPeer {
 
     @Override
     public void serializeBody(OutputStream stream) throws IOException {
-        writeInt(userId, stream);
+        writeLong(userId, stream);
     }
 
     @Override
@@ -58,7 +59,7 @@ public class TLPeerUser extends TLAbsPeer {
         return CONSTRUCTOR_ID;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
